Add files via upload
The excel file contains evaluation of floor altitude estimation for TUM Main Campus Library. The columns EG, 1st,..,5th in this file refer to each floor altitude of this building. Each column is a cluster from clustering result of the K-Means over the aggregated altitude data sets. If there are some columns without data, this means that the corresponding floor was not visited by the user in that  month.